% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageDataClass.R
\docType{data}
\name{PackageDataClass}
\alias{PackageDataClass}
\title{PackageDataClass}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PackageDataClass
}
\description{
A helper class
}
\keyword{datasets}
